requirements.txt
biopython
beautifulsoup4
datasets
matplotlib
numpy
openai
pandas
pydantic
scispacy
seaborn
spacy
scikit-learn
torch
transformers
